+++ src/Router.tsx
@@ -38,6 +38,7 @@ import { sep } from '@tauri-apps/api/path'
 import { paths } from 'lib/paths'
 import { IndexLoaderData, HomeLoaderData } from 'lib/types'
 import { fileSystemManager } from 'lang/std/fileSystemManager'
+import LspProvider from 'components/LspProvider'
 
 export const BROWSER_FILE_NAME = 'new'
 
@@ -52,7 +53,9 @@ const addGlobalContextToElements = (
           ...route,
           element: (
             <CommandBarProvider>
-              <SettingsAuthProvider>{route.element}</SettingsAuthProvider>
+              <SettingsAuthProvider>
+                <LspProvider>{route.element}</LspProvider>
+              </SettingsAuthProvider>
             </CommandBarProvider>
           ),
         }
